perf: Migrate Barkeeper customization gump to StaticGump
Splits BarkeeperGump (DynamicGump) into two StaticGump<T> variants
selected by body type — Human (modifiable appearance) and NonHuman
(no appearance/gender controls). Each variant gets its own cached
static layout via a CRTP base; dynamic per-instance text (rumor
messages, keywords, tip message) is filled via slot placeholders in
BuildStrings.

Moves PlayerBarkeeper, BarkeeperGump, and BarkeeperTitleGump into a
dedicated Mobiles/Vendors/Barkeeper/ folder.

Pulls the Back button on the appearance-categories page out of the
ModifyAppearance branch so non-human barkeepers no longer hit a
dead end on that page.
